32-bit build fix
git-svn-id: svn+ssh://market.scs.stanford.edu/home/u1/asbestos/repo/jos64/trunk@6746 927ebb28-5305-0410-b3ca-c4b640df69b7

+++ kern/kern/pagetree.c
@@ -101,7 +101,7 @@ pagetree_invalidate_ro_segments(void *pg)
 	uintptr_t parentip = (uintptr_t) pe->parent;
 	if (parentip & 1) {
 	    /* Points to an indirect page */
-	    void *parent = (void *) (parentip & UINT64(~0));
+	    void *parent = (void *) (uintptr_t) (parentip & UINT64(~0));
 	    pagetree_invalidate_ro_segments(parent);
 	    continue;
 	}
